Add rest of integration tests

+++ my_api/app/controllers/account_controller.rb
@@ -1,19 +1,44 @@
 class AccountController < ApplicationController
   def create
-    @user = User.find(params[:user_id])
+    begin
+      @user = User.find(params[:user_id])
+    rescue ActiveRecord::RecordNotFound
+      render json: 'User not found', status: :not_found
+      return
+    end
+
     @account = @user.accounts.create(number: params[:number])
 
     if @account.errors.any?
-      render json: @account.errors
+      render json: @account.errors, status: :bad_request
     else
-      render json: @account
+      render json: @account, status: :created
+    end
+  end
+
+  def update  
+    account = Account.find_by(number: params[:number])
+
+    if account.nil?
+      render json: 'Account not found', status: :not_found
+      return
+    end
+   
+
+    if account.update(account_params_edit)
+      render json: account, status: :ok
+    else
+      render json: 'Error updating user', status: :bad_request
     end
   end
 
   def withdraw
     account = Account.find_by(number: params[:number])
 
-    render json: 'Could not find account' if account.nil?
+    if account.nil?
+      render json: 'Account not found', status: :not_found
+      return
+    end
 
     withdraw_amount = params[:amount].to_f
 
@@ -21,7 +46,7 @@ class AccountController < ApplicationController
       render json: account
     else
       render json: "You don't have the required balance for this withdrawl.
-                    Balance: #{account.balance} Withdrawl: #{withdraw_amount}"
+                    Balance: #{account.balance} Withdrawl: #{withdraw_amount}", status: :bad_request
     end
   end
 
@@ -29,7 +54,7 @@ class AccountController < ApplicationController
     account = Account.find_by(number: params[:number])
 
     if account.nil?
-      render json: 'Could not find account'
+      render json: 'Could not find account', status: :not_found
       return
     end
 
@@ -47,10 +72,16 @@ class AccountController < ApplicationController
 
   def destroy
     account = Account.find_by(number: params[:number])
+
+    if account.nil?
+      render json: 'Account not found', status: :not_found
+      return
+    end
+
     if account.destroy
       render json: 'Destroyed'
     else
-      render json: 'Error destroying account'
+      render json: 'Error destroying account', status: :bad_request
     end
   end
 
@@ -61,7 +92,7 @@ class AccountController < ApplicationController
                                       params[:sender_id],
                                       params[:receiver_id])
     rescue => e
-      render json: e.message
+      render json: e.message, status: :bad_request
       return
     end
     render json: make_array_transfer(acc_receiver, acc_sender)
@@ -77,6 +108,10 @@ class AccountController < ApplicationController
      receiver: [account_receiver,
                 sent_transfers: account_receiver.sent_transfers,
                 received_transfers: account_receiver.received_transfers]]
+  end
+
+  def account_params_edit
+    params.permit(:balance)
   end
 
 end
